requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
coverage==7.0.1
Flask==2.2.2
Flask-Cors==1.10.3
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
nose==1.3.7
pinocchio==0.4.3
protobuf==3.20.1
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
zipp==3.11.0
